YoinkT2D/requirements.txt
numpy==2.2.4
Pillow==11.1.0
unitypy==1.21.2
